+++ CHANGELOG.rst
@@ -110,6 +110,6 @@ Changelog
 -----------------------------------------
 
 * Fixed task sync bug with missing destination tasks.
-* Implemented Habitica sync task
 * HabiticaService now supports a task filter on get_tasks
-
+* Implemented Habitica sync task
+* Implemented HabiticaTaskService